Reimplement Absorption using new mana ability system
